feat!: migrate to new calldata
- add parser for ts
- add encoder for ts
- bring python implementation from genvm
- patch backend for working with binary calldata

+++ tests/integration/contract_examples/mocks/llm_erc20_get_contract_schema_for_code.py
@@ -4,13 +4,13 @@ llm_erc20_contract_schema = {
     "result": {
         "abi": [
             {
-                "inputs": [{"name": "total_supply", "type": "uint256"}],
+                "inputs": [{"name": "total_supply", "type": "int"}],
                 "type": "constructor",
             },
             {
                 "inputs": [{"name": "address", "type": "string"}],
                 "name": "get_balance_of",
-                "outputs": [{"name": "", "type": "uint256"}],
+                "outputs": [{"name": "", "type": "int"}],
                 "type": "function",
             },
             {
@@ -21,7 +21,7 @@ llm_erc20_contract_schema = {
             },
             {
                 "inputs": [
-                    {"name": "amount", "type": "uint256"},
+                    {"name": "amount", "type": "int"},
                     {"name": "to_address", "type": "string"},
                 ],
                 "name": "transfer",
